pyproject.toml
[project]
name = "backend-portfolio"
version = "0.1.0"
description = "Backend for portfolio"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "cachetools==6.2.6",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.4",
    "deprecation==2.1.0",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.128.0",
    "fastapi-cli==0.0.20",
    "fastapi-cloud-cli==0.11.0",
    "fastar==0.8.0",
    "fsspec==2026.1.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "multidict==6.7.1",
    "packaging==26.0",
    "postgrest==2.27.2",
    "propcache==0.4.1",
    "psycopg2-binary==2.9.11",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.0",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyiceberg==0.10.0",
    "pyjwt==2.11.0",
    "pyparsing==3.3.2",
    "pyroaring==1.0.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "pyyaml==6.0.3",
    "realtime==2.27.2",
    "redis>=7.1.0",
    "requests==2.32.5",
    "rich==14.2.0",
    "rich-toolkit==0.17.1",
    "rignore==0.7.6",
    "sentry-sdk==2.50.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.45",
    "sqlmodel==0.0.31",
    "starlette==0.50.0",
    "storage3==2.27.2",
    "strenum==0.4.15",
    "strictyaml==1.7.3",
    "supabase==2.27.2",
    "supabase-auth==2.27.2",
    "supabase-functions==2.27.2",
    "tenacity==9.1.2",
    "typer==0.21.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.40.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
    "yarl==1.22.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
